+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### v1.7.2 (2021-10-28)
+- Pressing Enter while search bar is focused will now redirect to first result of local search ([#121](https://github.com/pawelmalak/flame/issues/121))
+- Use search bar shortcuts when it's not focused ([#124](https://github.com/pawelmalak/flame/issues/124))
+- Fixed bug with Weather API still logging with module being disabled ([#125](https://github.com/pawelmalak/flame/issues/125))
+- Added option to disable search bar autofocus ([#127](https://github.com/pawelmalak/flame/issues/127))
+
 ### v1.7.1 (2021-10-22)
 - Fixed search action not being triggered by Numpad Enter
 - Added option to change date formatting ([#92](https://github.com/pawelmalak/flame/issues/92))